fix(macos): split approval plan validation
