+++ apps/macos/Sources/OpenClaw/NodeMode/MacNodeRuntime.swift
@@ -1,5 +1,4 @@
 import AppKit
-import CryptoKit
 import Foundation
 import OpenClawIPC
 import OpenClawKit
@@ -928,7 +927,10 @@ extension MacNodeRuntime {
         }
         if approvalSource != nil || explicitApproval {
             guard let plan = params.systemRunPlan,
-                  Self.systemRunPlan(plan, matches: params, validatedCommand: validatedCommand)
+                  MacSystemRunApprovalPlanValidator.matches(
+                      plan,
+                      params: params,
+                      validatedCommand: validatedCommand)
             else {
                 let message = approvalSource != nil
                     ? "approvalSource requires matching systemRunPlan"
@@ -940,7 +942,7 @@ extension MacNodeRuntime {
         let delayedPolicySnapshot: ExecApprovalPolicySnapshot?
         if carriesDelayedAuthority {
             if let operand = params.systemRunPlan?.mutableFileOperand,
-               !Self.revalidateMutableFileOperand(
+               !MacSystemRunApprovalPlanValidator.revalidateMutableFileOperand(
                    operand,
                    command: validatedCommand.command,
                    cwd: params.cwd)
@@ -1037,78 +1039,6 @@ extension MacNodeRuntime {
             delayedPolicySnapshot: delayedPolicySnapshot,
             sessionKey: sessionKey,
             runId: runId))
-    }
-
-    private static func systemRunPlan(
-        _ plan: OpenClawSystemRunApprovalPlan,
-        matches params: OpenClawSystemRunParams,
-        validatedCommand: ExecHostValidatedRequest) -> Bool
-    {
-        // Delayed authority is signed for one normalized request. Match the
-        // same fields as the Node host before trusting its policy snapshot.
-        plan.argv == validatedCommand.command &&
-            plan.commandText == validatedCommand.displayCommand &&
-            self.normalizedSystemRunPlanString(plan.cwd) == self.normalizedSystemRunPlanString(params.cwd) &&
-            self.normalizedSystemRunPlanString(plan.agentId) == self.normalizedSystemRunPlanString(params.agentId) &&
-            self.normalizedSystemRunPlanString(plan.sessionKey) ==
-            self.normalizedSystemRunPlanString(params.sessionKey)
-    }
-
-    private static func normalizedSystemRunPlanString(_ value: String?) -> String? {
-        guard let value else { return nil }
-        let scalars = value.unicodeScalars
-        var start = scalars.startIndex
-        while start != scalars.endIndex, self.isECMAScriptTrimScalar(scalars[start]) {
-            start = scalars.index(after: start)
-        }
-        var end = scalars.endIndex
-        while end != start {
-            let previous = scalars.index(before: end)
-            guard self.isECMAScriptTrimScalar(scalars[previous]) else { break }
-            end = previous
-        }
-        let trimmed = String(scalars[start..<end])
-        return trimmed.isEmpty ? nil : trimmed
-    }
-
-    private static func revalidateMutableFileOperand(
-        _ operand: OpenClawSystemRunApprovalFileOperand,
-        command: [String],
-        cwd: String?) -> Bool
-    {
-        guard operand.argvIndex >= 0,
-              operand.argvIndex < command.count,
-              let rawPath = self.normalizedSystemRunPlanString(command[operand.argvIndex])
-        else { return false }
-
-        let basePath = self.normalizedSystemRunPlanString(cwd) ?? FileManager.default.currentDirectoryPath
-        let resolvedURL = URL(fileURLWithPath: rawPath, relativeTo: URL(fileURLWithPath: basePath, isDirectory: true))
-            .standardizedFileURL
-            .resolvingSymlinksInPath()
-        guard resolvedURL.path == operand.path,
-              let data = try? Data(contentsOf: resolvedURL, options: .mappedIfSafe)
-        else { return false }
-        let digest = SHA256.hash(data: data).map { String(format: "%02x", $0) }.joined()
-        return digest == operand.sha256
-    }
-
-    private static func isECMAScriptTrimScalar(_ scalar: Unicode.Scalar) -> Bool {
-        switch scalar.value {
-        case 0x0009...0x000D,
-             0x0020,
-             0x00A0,
-             0x1680,
-             0x2000...0x200A,
-             0x2028,
-             0x2029,
-             0x202F,
-             0x205F,
-             0x3000,
-             0xFEFF:
-            true
-        default:
-            false
-        }
     }
 
     private func handleSystemWhich(_ req: BridgeInvokeRequest) async throws -> BridgeInvokeResponse {
